requirements.txt
fastapi
uvicorn
yt-dlp
pydub
openai
python-dotenv
langchain
langchain-openai
langchain-community
langchain-core
streamlit

